requirements.txt
wheel
tinkoff-investments
python-dotenv
protobuf3
ta
sqlalchemy
torch
matplotlib
sklearn
pyqt6
